==> scripts/2.1_phyloseq.R <==
# Author : Jonathan Rondeau-Leclaire 2025

##############
# Contents ####
################

# 0.1. Load functions
# 0.2. Setup / metadata parsing
# 1.1. Process 16S samples
# 1.2. Process ITS samples
# 1.3. Process trnL samples
# 2.1. Export ps objects 

###############################
# 0.1. Functions & packages ####
#################################

library(pacman)
p_load(mgx.tools, # devtools::install_github("jorondo1/mgx.tools")
       tidyverse, phyloseq, magrittr, decontam, Biostrings,
       readxl, decontam)
# Decontam functions:
source("https://github.com/jorondo1/misc_scripts/raw/refs/heads/main/decontam_functions.R")

# Parse DNA concentration xlsx files
parse_CERMO_xlsx <- function(files) {
  require(readxl, dplyr)
  files %>% 
    map(read_xlsx, range = "C13:D120", # CERMO-files specific
        col_names = c('sample_id', 'concDNA')) %>% 
    list_rbind %>% 
    filter(!is.na(sample_id) 
           & !concDNA %in% c('n/a', 'Too Low')) %>% 
    mutate(sample_id = sub("_[^_]*$", "", sample_id),
           concDNA = as.numeric(concDNA))
}

# Add sequencing depth and dna concentration to sample metadata
add_seq_depth <- function(seqtab, meta, dnaConc) {
  require(tibble)
  meta_subset <- meta %>% 
    left_join(dnaConc, by = 'sample_id') %>% # adds seq depth column
    column_to_rownames('sample_id') %>% 
    .[rownames(seqtab),]
  
  seqtab %>% rowSums %>% 
    data.frame(seqDepth = .) %>% 
    cbind(meta_subset, .) 
}



######################
# 0.2. Parse metadata ###
########################

urbanbio.path <- '~/Desktop/ip34/urbanBio'
dna.path <- file.path(urbanbio.path,"data/metadata")

# Metadata
meta <- read_delim(file.path(urbanbio.path,"data/metadata/metadata_2022_samples_final.csv")) 

meta %>% filter(!is.na(time)) %>% 
  ggplot(aes(x = date, y= city, colour = time)) + geom_jitter(width = 0, height = 0.3) + theme_light()

# Weather data
weather <- Sys.glob(file.path(urbanbio.path,"data/metadata/meteo_*.csv")) %>% 
  map_dfr(read_delim, locale = locale(decimal_mark = ","),show_col_types = FALSE) %>% 
  mutate(city = case_when(
    `Nom de la Station` == 'BEAUPORT' ~ 'Québec',
    `Nom de la Station` == 'SHERBROOKE' ~ 'Sherbrooke',
    `Nom de la Station` == 'MCTAVISH' ~ 'Montréal'
  )) %>% 
  transmute(city = city,
            temp_moy = `Temp moy.(°C)`,
            precip = `Précip. tot. (mm)`,
            date = `Date/Heure`)

# Factors and weather data
meta %<>%
  mutate(across(where(is.character), as.factor)) %>% 
  left_join(weather, by = c('date', 'city')) %>% # Add precipitations & temperature data
  mutate(
    city = recode_factor(city, !!!c(
      'Montréal' = 'Montreal', 
      'Québec' = 'Quebec', 
      'Sherbrooke'='Sherbrooke')))

# Non-control samples
meta_samples <- meta %>% 
  filter(time != "None" & control == 'FALSE')

sample.names <- meta_samples$sample_id

# Controls
meta_ctrl <- meta %>% 
  filter(time =="None" | control=='TRUE') %>% 
  mutate(sample_id = case_when(
    sample_id == "Contrôle-blanc-12h00-PM" ~ "Controle-blanc-12h00-PM",
    TRUE ~ sample_id))

ctrl.names <- meta_ctrl$sample_id

####################
# 1.1. 16S samples ####
######################

path_16S <- file.path(urbanbio.path,'data/16S')
taxa_16S_genus <- read_rds(file.path(path_16S, '4_taxonomy/taxonomy.RDS'))
taxa_16S_species <- read_rds(file.path(path_16S, '4_taxonomy/taxonomy_species.RDS'))
seqtab_16S <- read_rds(file.path(path_16S, '4_taxonomy/seqtab.RDS'))

# Somehow assignSpecies drops all ranks except genus... fix 
Species_16S <- taxa_16S_species[,2]
names(Species_16S) <- rownames(taxa_16S_species)
taxa_16S <- cbind(taxa_16S_genus, Species_16S) %>% 
  data.frame %>% 
  mutate(Species = Species_16S, .keep = 'unused')
taxa_16S[is.na(taxa_16S)] <- 'Unclassified' 

# Replace Methylorubrum with Methylobacterium
taxa_16S[which(taxa_16S$Genus=='Methylorubrum'),"Genus"] <- "Methylobacterium"
which(taxa_16S$Genus=="Methylobacterium")

# Keep samples with metadata info
seqtab_16S_sam <- mgx.tools::subset_samples(seqtab_16S, sample.names)
seqtab_16S_ctrl <- mgx.tools::subset_samples(seqtab_16S, ctrl.names)

# Subset ASVs
taxa_16S_sam <- mgx.tools::subset_tax_table(taxa_16S, seqtab_16S_sam, 50) 
taxa_16S_ctrl <- mgx.tools::subset_tax_table(taxa_16S, seqtab_16S_ctrl, 50) 

# Seq count distribution
mgx.tools::viz_seqdepth(seqtab_16S_sam)

# Remove near-empty samples
seqtab_16S_sam_filt <- remove_ultra_rare(seqtab_16S_sam, taxa_16S_sam, 9000)
seqtab_16S_ctrl_filt <- remove_ultra_rare(seqtab_16S_ctrl, taxa_16S_ctrl, 10)
dim(seqtab_16S_sam); dim(seqtab_16S_sam_filt); dim(taxa_16S_sam); sum(seqtab_16S_sam_filt)/sum(seqtab_16S_sam)
dim(seqtab_16S_ctrl); dim(seqtab_16S_ctrl_filt); dim(taxa_16S_ctrl)

# Add sequencing effort and dna concentration to metadata
dna_16S <- Sys.glob(file.path(dna.path,'CERMO_*16s*.xlsx')) %>% parse_CERMO_xlsx()
meta_samples_16S <- add_seq_depth(seqtab_16S_sam_filt, meta_samples, dna_16S)
meta_ctrl_16S <- add_seq_depth(seqtab_16S_ctrl_filt, meta_ctrl, dna_16S)

# Find contaminants
contam_freq_16S <- decontaminate(seqtab_16S_sam_filt, meta_samples_16S, 'concDNA')
contam_freq_16S$p # look at contaminants correlation with concDNA

# Phyloseq object
ps_16S <- phyloseq(
  tax_table(taxa_16S_sam),
  otu_table(seqtab_16S_sam_filt, taxa_are_rows = FALSE),
  sample_data(meta_samples_16S)
) %>% prune_contam(contam_freq_16S$decontam) # decontam

ps_16S_ctrl <- phyloseq(
  tax_table(taxa_16S_ctrl),
  otu_table(seqtab_16S_ctrl_filt, taxa_are_rows = FALSE),
  sample_data(meta_ctrl_16S)
)

####################
# 1.2. ITS samples ####
######################

path_ITS <- file.path(urbanbio.path,'data/ITS')
taxa_ITS <- read_rds(file.path(path_ITS, '4_taxonomy/taxonomy.RDS'))
seqtab_ITS <- read_rds(file.path(path_ITS, '4_taxonomy/seqtab.RDS'))

taxa_ITS[is.na(taxa_ITS)] <- 'Unclassified' 

# Keep samples with metadata info
seqtab_ITS_sam <- mgx.tools::subset_samples(seqtab_ITS, sample.names)
seqtab_ITS_ctrl <- mgx.tools::subset_samples(seqtab_ITS, ctrl.names)

# Subset ASVs
taxa_ITS_sam <- mgx.tools::subset_tax_table(taxa_ITS, seqtab_ITS_sam, 50)
taxa_ITS_ctrl <- mgx.tools::subset_tax_table(taxa_ITS, seqtab_ITS_ctrl, 10)

# Seq count distribution
mgx.tools::viz_seqdepth(seqtab_ITS_sam)
mgx.tools::viz_seqdepth(seqtab_ITS_ctrl)

# Remove near-empty samples
seqtab_ITS_sam_filt <- remove_ultra_rare(seqtab_ITS_sam, taxa_ITS_sam, 2500)
seqtab_ITS_ctrl_filt <- remove_ultra_rare(seqtab_ITS_ctrl, taxa_ITS_ctrl, 0)

dim(seqtab_ITS_sam); dim(seqtab_ITS_sam_filt); dim(taxa_ITS_sam); sum(seqtab_ITS_sam_filt)/sum(seqtab_ITS_sam)
dim(seqtab_ITS_ctrl); dim(seqtab_ITS_ctrl_filt); dim(taxa_ITS_ctrl)

# Add sequencing effort and dna concentration to metadata
dna_ITS <- Sys.glob(file.path(dna.path,'CERMO_*ITS*.xlsx')) %>% parse_CERMO_xlsx()
meta_samples_ITS <- add_seq_depth(seqtab_ITS_sam_filt, meta_samples, dna_ITS)
meta_ctrl_ITS <- add_seq_depth(seqtab_ITS_ctrl_filt, meta_ctrl, dna_ITS)

# Phyloseq objects
ps_ITS <- phyloseq(
  tax_table(taxa_ITS_sam),
  otu_table(seqtab_ITS_sam_filt, taxa_are_rows = FALSE),
  sample_data(meta_samples_ITS)
  ) # %>% prune_contam(contam_freq_ITS$decontam) # a highly prevalent plant pathogen ...

ps_ITS_ctrl <- phyloseq(
  tax_table(taxa_ITS_ctrl),
  otu_table(seqtab_ITS_ctrl_filt, taxa_are_rows = FALSE),
  sample_data(meta_ctrl_ITS)
)

#######################
# 1.3. trnL samples ####
#########################

path_trnL <- file.path(urbanbio.path,'data/trnL')
taxa_trnL <- read_rds(file.path(path_trnL, '4_taxonomy_E22_100_trunc275/taxonomy.RDS'))
seqtab_trnL <- read_rds(file.path(path_trnL, '4_taxonomy_E22_100_trunc275/seqtab.RDS'))
taxa_trnL[is.na(taxa_trnL)] <- 'Unclassified' 
taxa_trnL <- taxa_trnL[which(taxa_trnL[,1] != 'Bacteria'),] # Remove cyanobacteria

# Keep samples with metadata info
seqtab_trnL_sam <- mgx.tools::subset_samples(seqtab_trnL, sample.names)
seqtab_trnL_ctrl <- mgx.tools::subset_samples(seqtab_trnL, ctrl.names)

# Subset ASVs
taxa_trnL_sam <- mgx.tools::subset_tax_table(taxa_trnL, seqtab_trnL_sam, min_seq = 50)
taxa_trnL_ctrl <- mgx.tools::subset_tax_table(taxa_trnL, seqtab_trnL_ctrl, min_seq = 50)

# Seq count distribution
mgx.tools::viz_seqdepth(seqtab_trnL_sam)

# Remove near-empty samples
seqtab_trnL_sam_filt <- remove_ultra_rare(seqtab_trnL_sam, taxa_trnL_sam, 2500)
seqtab_trnL_ctrl_filt <- remove_ultra_rare(seqtab_trnL_ctrl, taxa_trnL_ctrl, 2500)

dim(seqtab_trnL_sam); dim(seqtab_trnL_sam_filt); dim(taxa_trnL_sam); sum(seqtab_trnL_sam_filt)/sum(seqtab_trnL_sam)
dim(seqtab_trnL_ctrl); dim(seqtab_trnL_ctrl_filt); dim(taxa_trnL_ctrl)

# Add sequencing effort and dna concentration to metadata
dna_trnL <- Sys.glob(file.path(dna.path,'CERMO_*trnL*.xlsx')) %>% parse_CERMO_xlsx()
meta_samples_trnL <- add_seq_depth(seqtab_trnL_sam_filt, meta_samples, dna_trnL)
meta_ctrl_trnL <- add_seq_depth(seqtab_trnL_ctrl_filt, meta_ctrl, dna_trnL)

# Phyloseq objects
ps_trnL <- phyloseq(
  tax_table(taxa_trnL_sam),
  otu_table(seqtab_trnL_sam_filt, taxa_are_rows = FALSE),
  sample_data(meta_samples_trnL)
)

ps_trnL_ctrl <- phyloseq(
  tax_table(taxa_trnL_ctrl),
  otu_table(seqtab_trnL_ctrl_filt, taxa_are_rows = FALSE),
  sample_data(meta_ctrl_trnL)
) # %>% prune_contam(contam_freq_trnL$decontam) # birch?!

#######################"#######
# 1.4. trnL passive samples ####
#################################

# Passive 
meta_passive <- read_xlsx(file.path(urbanbio.path,"data/metadata/metadata_vaseline_sequencing.xlsx")) %>% 
  mutate(Sample_ID = str_replace_all(Sample_ID, "_", "-"),
         Start_date = parse_date(Start_date, c("%d-%m-%Y")),
         End_date = parse_date(End_date, c("%d-%m-%Y"))) %>% 
  separate_wider_delim(Sample_ID, delim = '-', 
                       names = c('Year', 'City', 'site_id' , 'TX'),
                       cols_remove= FALSE)
  
path_trnl_p <- file.path(urbanbio.path,'data/trnL_passive')
taxa_trnl_p <- read_rds(file.path(path_trnl_p, '4_taxonomy/taxonomy.RDS'))
seqtab_trnL_p <- read_rds(file.path(path_trnl_p, '4_taxonomy/seqtab.RDS'))
taxa_trnl_p[is.na(taxa_trnl_p)] <- 'Unclassified' 
taxa_trnl_p <- taxa_trnl_p[which(taxa_trnl_p[,1] != 'Bacteria'),] # Remove cyanobacteria

# Keep samples with metadata info
seqtab_trnL_sam_p <- mgx.tools::subset_samples(seqtab_trnL_p, meta_passive$Sample_ID)

# Subset ASVs
taxa_trnL_sam_p <- mgx.tools::subset_tax_table(taxa_trnl_p, seqtab_trnL_sam_p, min_seq = 50)

# Seq count distribution
mgx.tools::viz_seqdepth(seqtab_trnL_sam_p)

# Remove near-empty samples
seqtab_trnL_sam_filt_p <- remove_ultra_rare(seqtab_trnL_sam_p, taxa_trnL_sam_p, 500)

dim(seqtab_trnL_sam_p); dim(seqtab_trnL_sam_filt_p); dim(taxa_trnL_sam_p); sum(seqtab_trnL_sam_filt_p)/sum(seqtab_trnL_sam_p)

meta_passive_subset <- meta_passive %>% 
  filter(Sample_ID %in% rownames(seqtab_trnL_sam_filt_p)) %>% 
  column_to_rownames('Sample_ID')
  
# Phyloseq objects
ps_trnL_p <- phyloseq(
  tax_table(taxa_trnL_sam_p),
  otu_table(seqtab_trnL_sam_filt_p, taxa_are_rows = FALSE),
  sample_data(meta_passive_subset)
)

#####################
# 2.1. Export data ###
#######################
saveRDS(ps_trnL_p, 'data/ps_passive.rds')

ps.ls <- list()
ps.ls[["BACT"]] <- ps_16S
ps.ls[["FUNG"]] <- ps_ITS
ps.ls[["PLAN"]] <- ps_trnL
saveRDS(ps.ls, 'data/ps.ls.rds')

ps_ctrl.ls <- list()
ps_ctrl.ls[["BACT"]] <- ps_16S_ctrl
ps_ctrl.ls[["FUNG"]] <- ps_ITS_ctrl
ps_ctrl.ls[["PLAN"]] <- ps_trnL_ctrl
saveRDS(ps_ctrl.ls, 'data/ps_ctrl.ls.rds')

# Also save raw seqtab, taxtable, metadata
seqtab.ls <- list()
seqtab.ls[['BACT']] <- seqtab_16S_sam_filt
seqtab.ls[['FUNG']] <- seqtab_ITS_sam_filt
seqtab.ls[['PLAN']] <- seqtab_trnL_sam_filt
saveRDS(seqtab.ls, 'data/seqtab.ls.rds')

taxtab.ls <- list()
taxtab.ls[['BACT']] <- taxa_16S_sam
taxtab.ls[['FUNG']] <- taxa_ITS_sam
taxtab.ls[['PLAN']] <- taxa_trnL_sam
saveRDS(taxtab.ls, 'data/taxtab.ls.rds')

# Export asvs as fasta
asv_to_fasta(seqtab_16S_sam_filt, file.path(path_16S, '4_taxonomy/asv.fa'))
asv_to_fasta(seqtab_ITS_sam_filt, file.path(path_ITS, '4_taxonomy/asv.fa'))
asv_to_fasta(seqtab_trnL_sam_filt, file.path(path_trnL, '4_taxonomy_E22_100_trunc275/asv.fa'))


# Set browser
# Sys.setenv(CHROMOTE_CHROME = "/Applications/Brave Browser.app/Contents/MacOS/Brave Browser")
